requirements.txt
streamlit==1.38.0
openpyxl==3.1.5
reportlab==4.2.2
cryptography==42.0.8
python-dotenv==1.0.1
requests==2.31.0
watchdog==4.0.1
